+++ src/components/index.js
@@ -1,0 +1,27 @@
+import AmountCounter from './AmountCounter/AmountCounter';
+import Button from './Button/Button';
+import CheckBox from './CheckBox/CheckBox';
+import Header from './Header/Header';
+import Modal from './Modal/Modal';
+import SuccessAddedModal from './Modal/SuccessAddedModal/SuccessAddedModal';
+import NavBar from './NavBar/NavBar';
+import OrderContainer from './OrderContainer/OrderContainer';
+import PaymentInfoBox from './PaymentInfoBox/PaymentInfoBox';
+import ColumnProductItem from './ProductItem/ColumnProductItem/ColumnProductItem';
+import RowProductItem from './ProductItem/RowProductItem/RowProductItem';
+import ShoppingCartIcon from './ShoppingCartIcon/ShoppingCartIcon';
+
+export {
+  Button,
+  CheckBox,
+  Header,
+  Modal,
+  SuccessAddedModal,
+  NavBar,
+  OrderContainer,
+  PaymentInfoBox,
+  RowProductItem,
+  ColumnProductItem,
+  ShoppingCartIcon,
+  AmountCounter,
+};
