+++ src/pages/index.js
@@ -1,0 +1,6 @@
+import OrderCheckoutPage from './OrderCheckoutPage/OrderCheckoutPage';
+import OrderListPage from './OrderListPage/OrderListPage';
+import ProductListPage from './ProductListPage/ProductListPage';
+import ShoppingCartPage from './ShoppingCartPage/ShoppingCartPage';
+
+export { OrderCheckoutPage, ProductListPage, ShoppingCartPage, OrderListPage };
